Fix indexing large smithy-build.json (#13)
Fixes an issue where a UTFDataFormatException would be thrown from
SmithyBuildConfigurationIndex when saving the json value. Based on the
DataOutput.writeUTF docs:

https://docs.oracle.com/javase/8/docs/api/java/io/DataOutput.html#writeUTF-java.lang.String-
this happens when the number of bytes to be written is > 65535, so if
you have a very large smithy-build.json it throws. For some reason this
can also cause intellij to just index in an infinite loop, eventually
running out of memory.

I fixed it by writing the value out as bytes, which also required
reading it in differently too otherwise it couldn't be deserialized
after being read in again. There may be a better way to do this though.

Tested with `runIde` and opening the project that was previously giving
me problems.